+++ mobile/android/android-components/components/browser/icons/src/main/java/mozilla/components/browser/icons/extension/WebAppManifest.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.browser.icons.extension
 
+import android.graphics.Color
 import android.os.Build
 import android.os.Build.VERSION.SDK_INT
 import mozilla.components.browser.icons.IconRequest
@@ -19,13 +20,25 @@ import mozilla.components.concept.engine.manifest.WebAppManifest.Icon.Purpose
 fun WebAppManifest.toIconRequest() = IconRequest(
     url = startUrl,
     size = if (SDK_INT >= Build.VERSION_CODES.O) LAUNCHER_ADAPTIVE else LAUNCHER,
-    resources = icons.mapNotNull { it.toIconResource() },
+    resources = icons
+        .filter { Purpose.MASKABLE in it.purpose || Purpose.ANY in it.purpose }
+        .map { it.toIconResource() },
     color = backgroundColor
 )
 
-private fun WebAppManifest.Icon.toIconResource(): IconRequest.Resource? {
-    if (Purpose.MASKABLE !in purpose && Purpose.ANY !in purpose) return null
+/**
+ * Creates an [IconRequest] for retrieving a monochrome icon specified in the manifest.
+ */
+fun WebAppManifest.toMonochromeIconRequest() = IconRequest(
+    url = startUrl,
+    size = IconRequest.Size.DEFAULT,
+    resources = icons
+        .filter { Purpose.MONOCHROME in it.purpose }
+        .map { it.toIconResource() },
+    color = Color.WHITE
+)
 
+private fun WebAppManifest.Icon.toIconResource(): IconRequest.Resource {
     return IconRequest.Resource(
         url = src,
         type = MANIFEST_ICON,