+++ mobile/android/android-components/components/feature/pwa/src/main/java/mozilla/components/feature/pwa/feature/WebAppSiteControlsFeature.kt
@@ -11,14 +11,23 @@ import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.drawable.Icon
 import android.os.Build
+import android.os.Build.VERSION.SDK_INT
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationCompat.BADGE_ICON_NONE
 import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.getSystemService
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.OnLifecycleEvent
+import androidx.lifecycle.lifecycleScope
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.launch
+import mozilla.components.browser.icons.BrowserIcons
+import mozilla.components.browser.icons.extension.toMonochromeIconRequest
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.feature.pwa.R
@@ -30,12 +39,14 @@ import mozilla.components.feature.session.SessionUseCases
  * @param manifest Web App Manifest reference used to populate the notification.
  * @param controlsBuilder Customizes the created notification.
  */
+@Suppress("LongParameterList")
 class WebAppSiteControlsFeature(
     private val applicationContext: Context,
     private val sessionManager: SessionManager,
     private val sessionId: String,
     private val manifest: WebAppManifest? = null,
-    private val controlsBuilder: SiteControlsBuilder = SiteControlsBuilder.Default()
+    private val controlsBuilder: SiteControlsBuilder = SiteControlsBuilder.Default(),
+    private val icons: BrowserIcons? = null
 ) : BroadcastReceiver(), LifecycleObserver {
 
     constructor(
@@ -44,30 +55,71 @@ class WebAppSiteControlsFeature(
         reloadUrlUseCase: SessionUseCases.ReloadUrlUseCase,
         sessionId: String,
         manifest: WebAppManifest? = null,
-        controlsBuilder: SiteControlsBuilder = SiteControlsBuilder.CopyAndRefresh(reloadUrlUseCase)
+        controlsBuilder: SiteControlsBuilder = SiteControlsBuilder.CopyAndRefresh(reloadUrlUseCase),
+        icons: BrowserIcons? = null
     ) : this(
         applicationContext,
         sessionManager,
         sessionId,
         manifest,
-        controlsBuilder
+        controlsBuilder,
+        icons
     )
 
+    private var notificationIcon: Deferred<mozilla.components.browser.icons.Icon>? = null
+
+    /**
+     * Starts loading the [notificationIcon] on create.
+     */
+    @OnLifecycleEvent(Lifecycle.Event.ON_CREATE)
+    fun onCreate() {
+        if (SDK_INT >= Build.VERSION_CODES.M && manifest != null && icons != null) {
+            val request = manifest.toMonochromeIconRequest()
+            if (request.resources.isNotEmpty()) {
+                notificationIcon = icons.loadIcon(request)
+            }
+        }
+    }
+
+    /**
+     * Displays a notification from the given [SiteControlsBuilder.buildNotification] that will be
+     * shown as long as the lifecycle is in the foreground. Registers this class as a broadcast
+     * receiver to receive events from the notification and call [SiteControlsBuilder.onReceiveBroadcast].
+     */
     @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
-    fun onResume() {
+    fun onResume(owner: LifecycleOwner) {
         val filter = controlsBuilder.getFilter()
         applicationContext.registerReceiver(this, filter)
 
-        NotificationManagerCompat.from(applicationContext)
-            .notify(NOTIFICATION_TAG, NOTIFICATION_ID, buildNotification())
+        val manager = NotificationManagerCompat.from(applicationContext)
+        val iconAsync = notificationIcon
+        if (iconAsync != null) {
+            owner.lifecycleScope.launch {
+                val bitmap = iconAsync.await().bitmap
+                manager.notify(NOTIFICATION_TAG, NOTIFICATION_ID, buildNotification(bitmap))
+            }
+        } else {
+            manager.notify(NOTIFICATION_TAG, NOTIFICATION_ID, buildNotification(null))
+        }
     }
 
+    /**
+     * Cancels the site controls notification and unregisters the broadcast receiver.
+     */
     @OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
     fun onPause() {
         applicationContext.unregisterReceiver(this)
 
         NotificationManagerCompat.from(applicationContext)
             .cancel(NOTIFICATION_TAG, NOTIFICATION_ID)
+    }
+
+    /**
+     * Cancels the [notificationIcon] loading job on destroy.
+     */
+    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
+    fun onDestroy() {
+        notificationIcon?.cancel()
     }
 
     /**
@@ -82,19 +134,29 @@ class WebAppSiteControlsFeature(
     /**
      * Build the notification with site controls to be displayed while the web app is active.
      */
-    private fun buildNotification(): Notification {
-        val channelId = ensureChannelExists()
-
-        return NotificationCompat.Builder(applicationContext, channelId)
-            .setSmallIcon(R.drawable.ic_pwa)
-            .setContentTitle(manifest?.name ?: manifest?.shortName)
-            .setBadgeIconType(BADGE_ICON_NONE)
-            .setColor(manifest?.themeColor ?: NotificationCompat.COLOR_DEFAULT)
-            .setPriority(NotificationCompat.PRIORITY_MIN)
-            .setShowWhen(false)
-            .setOngoing(true)
-            .also { controlsBuilder.buildNotification(applicationContext, it, channelId) }
-            .build()
+    private fun buildNotification(icon: Bitmap?): Notification {
+        val builder = if (SDK_INT >= Build.VERSION_CODES.O) {
+            val channelId = ensureChannelExists()
+            Notification.Builder(applicationContext, channelId).apply {
+                setBadgeIconType(BADGE_ICON_NONE)
+            }
+        } else {
+            @Suppress("Deprecation")
+            Notification.Builder(applicationContext).apply {
+                setPriority(NotificationCompat.PRIORITY_MIN)
+            }
+        }
+        if (icon != null && SDK_INT >= Build.VERSION_CODES.M) {
+            builder.setSmallIcon(Icon.createWithBitmap(icon))
+        } else {
+            builder.setSmallIcon(R.drawable.ic_pwa)
+        }
+        builder.setContentTitle(manifest?.name ?: manifest?.shortName)
+        builder.setColor(manifest?.themeColor ?: NotificationCompat.COLOR_DEFAULT)
+        builder.setShowWhen(false)
+        builder.setOngoing(true)
+        controlsBuilder.buildNotification(applicationContext, builder)
+        return builder.build()
     }
 
     /**
@@ -103,7 +165,7 @@ class WebAppSiteControlsFeature(
      * Returns the channel id to be used for notifications.
      */
     private fun ensureChannelExists(): String {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+        if (SDK_INT >= Build.VERSION_CODES.O) {
             val notificationManager: NotificationManager = applicationContext.getSystemService()!!
 
             val channel = NotificationChannel(